Add state index to ST type.
The ST type now takes the form `ST omega s a b`, where `omega` is `C
tau` or `T` as before, and `s` is a new state type. This is used like in
Haskell's ST monad, except it tracks *reads*, not escaping
references. That is, a `take` expression has type

    ST (C a) a a b

whereas an expression that doesn't read, like `x := 1`, has type

    forall s a b . ST (C ()) s a b

We can tell when an action doesn't emit because it will be polymorphic
in the stream out type, i.e., it will have an `ST` type where the stream
output type is quantified over. However, an action could take but throw
away the result and still end up polymorphic in the stream input type!
Therefore any primitive that reads from a stream sets the `ST`'s state
type variable to be the stream's input type.