Fix casts.
When casting, we always want to produce a value; castExp is now named
castVal to reflect this.

When casting, we may not yet know what one of the type sis, so we first
attempt to unify them. If that fails, then we know we have concrete
types to deal with and we go from there.