Use casts to being if branches to the same type.
This patch uses casts to bring the branches of an if expression to the
same type, as we do with binary operators. Unlike binary operators, the
branches of an if expression need not be values, so we factored out the
code that casts compiled expressiosn into unifyCompiledExpTypes.

The error message when expressiosn cannot be cast to a common type
should be better now; if we can't generate a cast, we attempt to unify,
so the error message is the same as when unification fails.