Only cast down integer constants.
Integer constants that are involved in, e.g., binary operations, can
always be cast "down" to another numeric type. This lets us write things
like 1 + x, where x is an int8, and get an int8 back instead of an
int32.

Before we allowed *all* constants to be cast down, but that is the wrong
thing to do for float constants!