Fix par compilation.
We must make a distinction between the end of of a control branch, i.e.,
in an if computation, end the end of a computation in a par, i.e., where
the done continuation should be called. We cannot terminate the left and
rigth sides of a par with a Pure return, because monadic bind will then
attempt to simultaneously bind both branches of the par to the next
computation in sequence after the par.