Properly instantiate type variables in Comps.
The Comp data type carries type, but when we compile the Comp, we don't
have access to the type environment in which the comp originated. In
particular, we don't know how type variables were instantiated at the
point the comp occurred. However, at the time we construct a Comp,
we *do* have access to the type environment, so now we substitute for
all type variables at Comp construction time.

This general solution also means we can get rid of the old hack we had
to use when constructing a take/takes Comp.