Clean up the MonadErr class.
Previously, there was no way to modify the entire error context, so it
was difficult to save the current state and restore it when, e.g.,
wrapping a base MonadErr in a continuation monad. localLocContext has
been repurposed to set the entire error context, and a new function,
withLocContext, adds just one piece of error context.