Keep par threads running at all times.
Instead of spinning up a new thread every time we need one for a par, we
spin them all up at initialization time and have them all wait on a
semaphore until they are needed. When a thread is done computing, it
waits on the semaphore until it is needed again.