Clean up the compiler pipeline.
The pipeline is now formatted to make it clearer that we run the linter
after each phase. We also now only run the linter after a phase if the
phase actually runs---before we were linting the result of a phase that
didn't run and was therefore just the identity on programs.