Clean up the "pureish" type story.
This patch started with my effort to modify the linter to check that
lifted computations, i.e., arguments to LiftC in the auto language, were
actually pureish. That required revamping when ST types were
instantiated. This has the side effect of giving us better information
about which computations are actually pureish since we lose the ability
to determine that after an ST type is instantiated.