Avoid shadowing when transforming to auto.
When converting to the auto language, we flatten all monadic
actions. For this to be sound, we need to make sure that bindings
introduced in a nested moandic action don't shadow variables, e.g.,

 { let x = 1
 ; y <- { let x = 2
        ; return x
        }
 ; return x
 }

/should not/ become

 { let x = 1
 ; let x = 2
 ; y <- return x
 ; return x
 }

Now we rename any bindings that are currently in scope to prevent
shadowing.

Interestingly, this bug didn't bite us on the test suite...