Add function for rewriting initial labels.
When inlining, we transform a step into a sequence of steps. However, if
the label of the step we transformed is referred to elsewhere, we need
to rewrite the inlined computation so it has the old initial label. Of
course we then need to rewrite the body of the inlined computation so
that it refers to the "old" label too.