Enforce order of side effects when dereferencing.
Previously, we were not ensuring that the current value of the
expression being dereferenced was being properly saved. Instead, the
actually dereferencing could be delayed. In particular, this happened
when a reference was represented as a plain old C variable.

Now we enforce the order of side effects that result from dereferencing
values. This also required fixing the way while loops are compiled; see
the note [Compiling While Loops].