Don't allow symbolic values when assigning.
When assigning to a ref, do not allow a symbolic value to be
assigned. The ref may go out of scope /after/ variables in the assigned
symbolic value go out of scope, leading to undefined variables.

The other option would be to check all symbolic heap values every time a
variable goes out of scope and remove those that contain the newly
out-of-scope variable, but that seems quite inefficient.