Avoid extraneous dereferencing.
Previously, we always stored every dereferenced value in an explicit
temporary. While the C compiler should be able to optimize these extra
temporaries away for scalar values, for arrays these copies manifest as
calls to memcpy, which the compiler surely cannot optimize away.

With this change, we only store a dereferenced value if the result of
the dereferenced value might be used after the ref is changed.

This fixes #8.