Add a ref flow analysis phase.
The reference flow analysis tracks what variables references flow to. We
use this information to void copying references immediately upon
dereference. Instead, we only perform the copy if the destination of the
dereference may be used after the reference is modified. This avoids a
large number of copies. If a value derived from a reference may be used
after the reference is modified, we call it tainted.

We were sort of doing this before, but the new analysis attaches taint
information to binders, making it easier to use. The new analysis
happens to be more accurate (and principled).