Limit declaration scope.
Previously, all variable declarations were placed at the top of the
enclosing function's body except for the special case of index variables
for pure loops. We now make a serious effort to minimize the scope of a
declaration. Hopefully this will make the C compiler's job easier and
will lead to less stack space being used.