Track left and right continuation labels.
Tracking just one continuation label was never correct, but it sort of
worked. Tracking both continuation labels allows us to fuse more
for/repeat loops, but this can lead to LoopC steps being inserted at the
end of, e.g., for bodies. We now handle that as well with mkUnloopComp.