Take LUTting into account during fusion blowup.
Fusion can result in code explosion, which we try to limit with a blowup
factor. However, code explosion is fine as long as the larger expression
will be LUTted---if fact, this is a major benefit of fusion. This patch
attempts to auto-LUT an expression that we would otherwise reject due to
code explosion, and if auto-LUTting is successful, we keep the fused
expression despite the blowup.