More read-write set fixes.
This batch of fixes is related to the fact that read sets are *may* read
sets but write sets must be precise. We were handling arrays correctly,
but not plain old variables. This bit us when we had an if where one
branch wrote a variable and the other did not.

There is also a small improvement to calculating read variables---we use
the bounded interval to figure out which part of an array to read, but
the precise interval to figure out what part to write. We used to always
use the precise interval, which could be worse than the bounded interval
if we, e.g., read from disjoint segments of an array.