Reset seen labels when collecting a loop head.
When collecting a loop head, we reset the set containing labels we have
already seen to the empty set. Otherwise we can end up with a loop head
that is outside the body of code currently being generated.

How could we encounter a previously encountered label? Sometimes when
merging loops, the last label we encounter on the left side of the loop
is the loop head...