Make sure the consumer always drains the queue.
The thread done flag is used to indicate both that one of the two
computations in the par terminated while computing a final result and
that the producer has run out of input. Therefore, we always ensure that
the consumer drains the queue even when the done flag is set.

This also matches the sequential semantics better, where the consumer
would consume all inmput before the producer could terminate.