Produce/consumer have different exit conditions.
Producer and consumer threads have different exit conditions.

The producer needs to exit if the consumer computed a final final value.

The consumer needs to exit if the producer terminated, which could be
because it ran out of input or because it computed a final value, *and*
if the queue has been drained.