Add consumer strategy for compiling threaded par.
The consumer strategy spins off a thread for the consumer of a
par. Since the consumer runs the par's continuation, this strategy ends
up running the entire rest of the computation in a new thread.

This thread only works for top-level pars, so that is the only time we
use it.