Rework use of labels in fusion pass.
The old fusion pass tried to be tricky about labels when unrolling loops. We
need to be careful about labels because that is how we track the program counter
and detect loops. Now we don't relabel unrolled statement; instead, we add dummy
statements with the appropriate label. This seems to be much less fragile than
our previous approach.